@@ -1,3 +1,3 @@
 module github.com/pirsch-analytics/pirsch-go-sdk
 
-go 1.15
+go 1.16
